Implemented decode, added instructions 1, 3 and 6

+++ chip8Emulator/include/memory.h
@@ -9,7 +9,7 @@ namespace chip8 {
 	public:
 		Memory();
 		~Memory();
-		uint8_t read(uint16_t address);
+		uint16_t read(uint16_t address);
 		void write(uint16_t address, uint8_t data);
 		void write(uint32_t address, const uint8_t* data, size_t dataSizeInBytes);
